Add `endpoint` parameter to allow users to mock out AWS APIs.
This new options allows developers to run this plugin against a mock
AWS server (like moto). For example, if the mock server is running
on localhost:5000, you can connect to the server (instead of the
production AWS server) with the following settings:

    <source>
      @type cloudwatch_logs
      tag debug.log
      endpoint http://localhost:5000
      ...
    </source>

This should come in very handy especially when tinkering the features
of the plugin.

Signed-off-by: Fujimoto Seiji <fujimoto@clear-code.com>