test: Migrate test cases to use driver.feed() instead of driver.emit()
The problem is that the test suite is still relying on the old API (e.g.
`OutputTestDriver.emit`) that has been deprecated since Fluentd v0.14.
As a consequence, it will raise 'undefined method' exceptions when the
version of the host server is Fluentd v1.0 or later.

This patch fixes the issue by migrating these test cases to v1.0 API.