Add an option to select the JSON library to use (JSON/Yajl)
This patch adds a configurable parameter which enables developers to
choose their preferable JSON library. This option should be useful to
mitigate the encoding-related errors, which are apparently caused by
the subtle implementation difference in the JSON libraries.

For now, it only supports the following two libraries: JSON (which is
the default choice) and Yajl, but it should be fairly trivial to add
other options here.